+++ client/src/three/scenes/constants.ts
@@ -15,6 +15,7 @@ export const structureTypeToBuildingType: Record<StructureType, BuildingType> = 
 };
 
 export const buildingModelPaths: Record<BuildingType, string> = {
+  [BuildingType.None]: "",
   [BuildingType.Bank]: "/models/buildings/bank.glb",
   [BuildingType.ArcheryRange]: "/models/buildings/archer_range.glb",
   [BuildingType.Barracks]: "/models/buildings/barracks.glb",
